requirements.txt
streamlit
psycopg2-binary
pgvector
sqlalchemy
langchain-postgres
langchain-huggingface
sentence-transformers
huggingface-hub
opencv-python-headless
Pillow
pandas
bcrypt
python-dotenv
qrcode[pil]
phonenumbers
surya-ocr
fuzzywuzzy
python-Levenshtein
openpyxl

